bit assembler v1. supports basic high-level-like syntax. more upgrades coming soon. (optimizations, standard library, etc)

+++ file.h
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-typedef uint8_t u8;
+typedef long long u8;
 typedef char   *string;
 #define bool  char
 #define false 0
